Add the initial implementation of the process kernel stack logic.  Not yet integrated into the main OS logic nor tested.

+++ arch/arm/include/armv7-a/irq.h
@@ -293,7 +293,7 @@ struct xcptcontext
    * stack in place.
    */
 
-  FAR uintptr_t *kstack;
+  FAR uint32_t *kstack;
 #endif
 #endif
 };
